Estimate the registration between two cartridge case scans using a cell-based or full scan comparison procedure

The comparison_cellBased function uses a cell-based comparison procedure implemented in the cmcR::comparison_allTogether(). The comparison_fullScan estimates the registration by comparing the full scans.

Usage

comparison_cellBased(
  reference,
  target,
  thetas = seq(-30, 30, by = 3),
  numCells = c(8, 8),
  maxMissingProp = 0.99,
  sideLengthMultiplier = 3,
  returnX3Ps = TRUE
)

comparison_fullScan(
  reference,
  target,
  thetas = seq(-30, 30, by = 3),
  returnX3Ps = TRUE
)

Arguments

reference

an x3p object

target

another x3p object

thetas

a numeric vector of arbitrary length containing rotation values to be considered in the registration procedure

numCells

a numeric vector of length 2 containing the number of cells to divide the c(rows,cols) of a cartridge case surface matrix into.

maxMissingProp

maximum proportion of missing values allowed for each cell/region.

sideLengthMultiplier

ratio between the target region and reference cell side lengths. For example, sideLengthMultiplier = 3 implies each region will be 9 times larger than its paired reference cell.

returnX3Ps

A Boolean to return the cellHeightValues and alignedTargetCells for each cell index. Note that setting this argument to TRUE is required to calculate the visual diagnostic features

Note

The distinction between the "reference" and "target" scans is arbitrary, yet necessary for keeping track of the comparison direction in which a particular observation was calculated.

See also

Examples

data("K013sA1","K013sA2")

compData_cellBased <- comparison_cellBased(reference = K013sA1,
                                           target = K013sA2,
                                           thetas = c(-3,0,3))

compData_cellBased
#> # A tibble: 234 x 12
#>    cellIndex     x     y fft_ccf pairwis~1 theta refMi~2 targM~3 joint~4 cellH~5
#>    <chr>     <dbl> <dbl>   <dbl>     <dbl> <dbl>   <dbl>   <dbl>   <dbl> <named>
#>  1 1, 2        -44    23   0.192     0.521    -3    2454   19377    2245 <x3p>  
#>  2 1, 3        -31    18   0.281     0.626    -3    1150   18225    1133 <x3p>  
#>  3 1, 4        -32    15   0.386     0.707    -3    1103   20221    1016 <x3p>  
#>  4 1, 5         30    -1   0.265     0.433    -3    1450   22308    1142 <x3p>  
#>  5 1, 6          0     2   0.265     0.647    -3    1751   23429    1552 <x3p>  
#>  6 1, 7        -50   -44   0.454     0.754    -3    2401   23848    2380 <x3p>  
#>  7 2, 1         -8     0   0.171     0.605    -3    2498   18498    2171 <x3p>  
#>  8 2, 2        -26    25   0.158     0.372    -3     378   15675     357 <x3p>  
#>  9 2, 3        -30    21   0.142     0.503    -3    1912   16268    1880 <x3p>  
#> 10 2, 7        -24   -26   0.352     0.642    -3    2086   23306    2000 <x3p>  
#> # ... with 224 more rows, 2 more variables: alignedTargetCell <named list>,
#> #   direction <chr>, and abbreviated variable names 1: pairwiseCompCor,
#> #   2: refMissingCount, 3: targMissingCount, 4: jointlyMissing,
#> #   5: cellHeightValues
#> # i Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

compData_fullScan <- comparison_fullScan(reference = K013sA1,
                                         target = K013sA2,
                                         thetas = c(-3,0,3))

compData_fullScan
#> # A tibble: 6 x 12
#>   cellIndex     x     y fft_ccf pairwise~1 theta refMi~2 targM~3 joint~4 cellH~5
#>   <chr>     <dbl> <dbl>   <dbl>      <dbl> <dbl>   <dbl>   <dbl>   <dbl> <named>
#> 1 1, 1        -14    -8  0.0955      0.140    -3  126310  159739  113854 <x3p>  
#> 2 1, 1         -8     0  0.137       0.187     0  126310  159717  116953 <x3p>  
#> 3 1, 1        -11    12  0.270       0.400     3  126310  159722  117849 <x3p>  
#> 4 1, 1         12   -12  0.272       0.410    -3  123397  167983  119276 <x3p>  
#> 5 1, 1          8     0  0.137       0.187     0  123397  167958  119212 <x3p>  
#> 6 1, 1         14     7  0.0950      0.138     3  123397  167970  116030 <x3p>  
#> # ... with 2 more variables: alignedTargetCell <named list>, direction <chr>,
#> #   and abbreviated variable names 1: pairwiseCompCor, 2: refMissingCount,
#> #   3: targMissingCount, 4: jointlyMissing, 5: cellHeightValues
#> # i Use `colnames()` to see all variable names