Calculate features based on visual diagnostic plots

The feature_visualDiagnostic_all() calculates the six visual diagnostic-based features in one call. Calculate each feature individually with feature_visualDiagnostic_neighborhoodSizeSummary() and feature_visualDiagnostic_scanDifferenceCor().

Usage

feature_visualDiagnostic_all(
  comparisonData,
  thresholdMultiplier = 1,
  id_cols = NULL
)

feature_visualDiagnostic_neighborhoodSizeSummary(
  cellHeightValues,
  alignedTargetCell,
  summaryFun = mean,
  thresholdMultiplier = 1,
  imputeVal = 1000
)

feature_visualDiagnostic_scanDifferenceCor(
  cellHeightValues,
  alignedTargetCell,
  thresholdMultiplier = 1,
  imputeVal = -1
)

Arguments

comparisonData

tibble such as one returned by the comparison_cellBased() or comparison_fullScan() functions that contains results from the cell-based or full scan comparison procedure

thresholdMultiplier

the default filtering threshold is equal to the standard deviation of the joint x3p1 and x3p2 surface matrix values. This argument can be used to scale this threshold

id_cols

column names in the comparisonData tibble that uniquely identify each observation. These are returned along with the computed features

cellHeightValues

list/tibble column of x3p objects containing a reference scan's cells (as returned by comparison_cellBased() or comparison_fullScan())

alignedTargetCell

list/tibble column of x3p objects containing a target scan's aligned cells (as returned by comparison_cellBased() or comparison_fullScan())

summaryFun

function that will be used to summarize the neighborhood sizes

imputeVal

value to return if the feature calculation results in a non-numeric (i.e., NA, NULL) value

Note

The feature_visualDiagnostic_all function can be used on comparison data from a full-scan or cell-based comparison. For a full-scan comparison, note that the standard deviation features will always be NA.

Examples

data("K013sA1","K013sA2")

compData_cellBased <- comparison_cellBased(reference = K013sA1,
                                           target = K013sA2,
                                           thetas = c(-3,0,3))

compData_cellBased %>%
  dplyr::group_by(cellIndex,direction) %>%
  feature_visualDiagnostic_all()
#> # A tibble: 78 x 8
#> # Groups:   cellIndex [39]
#>    cellIndex direction           neigh~1 neigh~2 neigh~3 neigh~4 diffe~5 diffe~6
#>    <chr>     <chr>                 <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 1, 2      reference vs. targ~    34.9    2.55    81.9    6.41   0.363  0.106 
#>  2 1, 2      target vs. referen~    83.7   36.0    126.    24.3    0.347  0.0571
#>  3 1, 3      reference vs. targ~    92.3   78.1    130.     8.79   0.282  0.0326
#>  4 1, 3      target vs. referen~    36.0    1.44   144.     8.20   0.199  0.0450
#>  5 1, 4      reference vs. targ~    64.0   16.0    123.    17.4    0.541  0.0249
#>  6 1, 4      target vs. referen~    51.5   10.5    108.    21.4    0.340  0.194 
#>  7 1, 5      reference vs. targ~    38.9    7.36    97.6    9.05   0.144  0.0830
#>  8 1, 5      target vs. referen~   102.    30.2    114.    NA      0.622  0.153 
#>  9 1, 6      reference vs. targ~    77.3   18.5    129.    13.9    0.494  0.0714
#> 10 1, 6      target vs. referen~    58.0   10.0    101.    10.2    0.204  0.126 
#> # ... with 68 more rows, and abbreviated variable names
#> #   1: neighborhoodSizeAve_ave, 2: neighborhoodSizeAve_sd,
#> #   3: neighborhoodSizeSD_ave, 4: neighborhoodSizeSD_sd, 5: differenceCor_ave,
#> #   6: differenceCor_sd
#> # i Use `print(n = ...)` to see more rows

compData_fullScan <- comparison_fullScan(reference = K013sA1,
                                         target = K013sA2,
                                         thetas = c(-3,0,3))

compData_fullScan %>%
  dplyr::group_by(direction) %>%
  feature_visualDiagnostic_all() %>%
  dplyr::select(-c(neighborhoodSizeAve_sd,
            neighborhoodSizeSD_sd,
            differenceCor_sd))
#> # A tibble: 2 x 4
#>   direction            neighborhoodSizeAve_ave neighborhoodSizeSD_ave differen~1
#>   <chr>                                  <dbl>                  <dbl>      <dbl>
#> 1 reference vs. target                    90.0                   936.    0.00721
#> 2 target vs. reference                    86.6                   913.    0.0113 
#> # ... with abbreviated variable name 1: differenceCor_ave